+++ components/screenstreaming/impl/src/main/java/com/flipperdevices/screenstreaming/impl/composable/ComposableStreamingScreen.kt
@@ -27,7 +27,6 @@ import com.flipperdevices.screenstreaming.impl.R
 import com.flipperdevices.screenstreaming.impl.composable.controls.ComposableFlipperControls
 import com.flipperdevices.screenstreaming.impl.composable.screen.ComposableFlipperScreenWithOptions
 import com.flipperdevices.screenstreaming.impl.model.FlipperLockState
-import com.flipperdevices.screenstreaming.impl.viewmodel.LockViewModel
 import com.flipperdevices.screenstreaming.impl.viewmodel.ScreenStreamingViewModel
 import com.flipperdevices.screenstreaming.impl.viewmodel.ScreenshotViewModel
 import kotlinx.collections.immutable.toImmutableList
@@ -39,13 +38,12 @@ import com.flipperdevices.core.ui.res.R as DesignSystem
 fun ComposableStreamingScreen(
     screenStreamingViewModel: ScreenStreamingViewModel,
     screenshotViewModel: ScreenshotViewModel,
-    lockViewModel: LockViewModel,
     modifier: Modifier = Modifier,
     onBack: () -> Unit
 ) {
     val flipperScreen by screenStreamingViewModel.getFlipperScreen().collectAsState()
     val buttons by screenStreamingViewModel.getFlipperButtons().collectAsState()
-    val lockState by lockViewModel.getLockState().collectAsState()
+    val lockState by screenStreamingViewModel.getLockState().collectAsState()
 
     Column(
         modifier = modifier
@@ -72,7 +70,7 @@ fun ComposableStreamingScreen(
                     when (it) {
                         FlipperLockState.NotInitialized -> {}
                         FlipperLockState.NotSupported -> showDeprecatedLockDialog = true
-                        is FlipperLockState.Ready -> lockViewModel.onChangeLock(!it.isLocked)
+                        is FlipperLockState.Ready -> screenStreamingViewModel.onChangeLock(!it.isLocked)
                     }
                 }
             }